requirements.txt
python-dotenv>=1.0.1
pydantic>=2.7
streamlit>=1.36
azure-ai-documentintelligence>=1.0.0
openai>=1.40.0
Pillow>=10.0.0